+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Refactor ToastManager and simplify state usage](https://github.com/multiversx/mx-sdk-dapp/pull/1603)
+
 ## [[5.2.3](https://github.com/multiversx/mx-sdk-dapp/pull/1601)] - 2025-09-29
 
 - [Fixed reading from state with immer current](https://github.com/multiversx/mx-sdk-dapp/pull/1600)