+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[v4.3.5](https://github.com/multiversx/mx-sdk-dapp/pull/1601)] - 2025-10-06
+
+- [Added ability to bypass sender validation for specific origins in `checkIsValidSender`](https://github.com/multiversx/mx-sdk-dapp/pull/1601)
+
 ## [[v4.3.4](https://github.com/multiversx/mx-sdk-dapp/pull/1600)] - 2025-09-27
 
 - [Fixed `useWalletConnectV2Login` logs out on `initiateLogin` when is logged in](https://github.com/multiversx/mx-sdk-dapp/pull/1600)